Fix up platform specific parts in test
