+++ test/src/androidMain/kotlin/io/realm/test/platform/RunLoopThread.kt
@@ -14,37 +14,41 @@
  * limitations under the License.
  */
 
-package io.realm.util
+package io.realm.test.platform
 
-import kotlinx.coroutines.CoroutineDispatcher
+import android.os.Handler
+import android.os.HandlerThread
+import io.realm.test.platform.RunLoopThread
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.android.asCoroutineDispatcher
 import kotlinx.coroutines.async
-import platform.CoreFoundation.CFRunLoopGetCurrent
-import platform.CoreFoundation.CFRunLoopRun
-import platform.CoreFoundation.CFRunLoopStop
+import java.util.concurrent.CountDownLatch
 import kotlin.coroutines.CoroutineContext
 
 actual class RunLoopThread : CoroutineScope {
 
-    private val dispatcher: CoroutineDispatcher by lazy { Main }
-    override val coroutineContext: CoroutineContext by lazy { dispatcher + exceptionHandler }
+    private val exit = CountDownLatch(1)
+
+    private val thread = HandlerThread("test-thread")
+    private val handler: Handler by lazy { thread.start(); Handler(thread.looper) }
+
+    override val coroutineContext: CoroutineContext by lazy { handler.asCoroutineDispatcher() + exceptionHandler }
 
     private var error: Throwable? = null
 
     val exceptionHandler = CoroutineExceptionHandler { _, exception ->
         error = exception
-        println("CoroutineExceptionHandler got $exception")
         terminate()
     }
 
     actual fun run(block: RunLoopThread.() -> Unit) {
         this.async { block(this@RunLoopThread) }
-        CFRunLoopRun()
+        exit.await()
         error?.let { throw it }
     }
 
     actual fun terminate() {
-        CFRunLoopStop(CFRunLoopGetCurrent())
+        exit.countDown()
     }
 }
