+++ test/src/androidTest/kotlin/io/realm/MemoryTests.kt
@@ -21,8 +21,8 @@ import android.os.Process
 import android.text.format.Formatter
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import io.realm.util.PlatformUtils
-import io.realm.util.PlatformUtils.triggerGC
+import io.realm.test.platform.PlatformUtils
+import io.realm.test.platform.PlatformUtils.triggerGC
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
