+++ test/src/androidTest/kotlin/io/realm/shared/ImportTests.kt
@@ -19,7 +19,7 @@ import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.RealmObject
 import io.realm.isManaged
-import io.realm.util.PlatformUtils
+import io.realm.test.platform.PlatformUtils
 import io.realm.util.TypeDescriptor.classifiers
 import io.realm.util.Utils.createRandomString
 import test.Sample
