+++ test/src/androidTest/kotlin/io/realm/shared/MutableRealmTests.kt
@@ -15,7 +15,7 @@
  */
 package io.realm
 
-import io.realm.util.PlatformUtils
+import io.realm.test.platform.PlatformUtils
 import io.realm.util.Utils.createRandomString
 import test.StringPropertyWithPrimaryKey
 import test.link.Child
