+++ test/src/androidTest/kotlin/io/realm/shared/NullabilityTests.kt
@@ -17,7 +17,7 @@ package io.realm.shared
 
 import io.realm.Realm
 import io.realm.RealmConfiguration
-import io.realm.util.PlatformUtils
+import io.realm.test.platform.PlatformUtils
 import io.realm.util.Utils.createRandomString
 import test.Nullability
 import kotlin.test.AfterTest
