+++ test/src/androidTest/kotlin/io/realm/shared/PrimaryKeyTests.kt
@@ -18,7 +18,7 @@ package io.realm.shared
 import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.RealmObject
-import io.realm.util.PlatformUtils
+import io.realm.test.platform.PlatformUtils
 import io.realm.util.TypeDescriptor.allPrimaryKeyFieldTypes
 import io.realm.util.TypeDescriptor.rType
 import io.realm.util.Utils.createRandomString
