+++ test/src/androidTest/kotlin/io/realm/shared/RealmResultsTests.kt
@@ -19,7 +19,7 @@ import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.RealmResults
 import io.realm.VersionId
-import io.realm.util.PlatformUtils
+import io.realm.test.platform.PlatformUtils
 import io.realm.util.Utils.createRandomString
 import test.link.Child
 import test.link.Parent
