+++ test/src/androidTest/kotlin/io/realm/shared/RealmTests.kt
@@ -19,8 +19,8 @@ import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.VersionId
 import io.realm.isManaged
-import io.realm.util.PlatformUtils
-import io.realm.util.PlatformUtils.triggerGC
+import io.realm.test.platform.PlatformUtils
+import io.realm.test.platform.PlatformUtils.triggerGC
 import io.realm.util.Utils.createRandomString
 import io.realm.version
 import kotlinx.coroutines.Job
