+++ test/src/androidTest/kotlin/io/realm/shared/SampleTests.kt
@@ -20,7 +20,7 @@ import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.RealmResults
 import io.realm.delete
-import io.realm.util.PlatformUtils
+import io.realm.test.platform.PlatformUtils
 import io.realm.util.Utils.createRandomString
 import test.Sample
 import kotlin.test.AfterTest
