+++ test/src/androidTest/kotlin/io/realm/shared/notifications/RealmListNotificationsTests.kt
@@ -3,7 +3,7 @@ package io.realm.shared.notifications
 import io.realm.NotificationTests
 import io.realm.Realm
 import io.realm.RealmConfiguration
-import io.realm.util.PlatformUtils
+import io.realm.test.platform.PlatformUtils
 import test.Sample
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
