+++ test/src/androidTest/kotlin/io/realm/shared/notifications/RealmNotificationsTests.kt
@@ -5,7 +5,7 @@ import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.VersionId
 import io.realm.internal.platform.runBlocking
-import io.realm.util.PlatformUtils
+import io.realm.test.platform.PlatformUtils
 import kotlinx.coroutines.async
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.collect
