+++ test/src/androidTest/kotlin/io/realm/shared/notifications/RealmResultsNotificationsTests.kt
@@ -5,7 +5,7 @@ import io.realm.NotificationTests
 import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.RealmResults
-import io.realm.util.PlatformUtils
+import io.realm.test.platform.PlatformUtils
 import kotlinx.coroutines.async
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.collect
