+++ test/src/macosTest/kotlin/io/realm/CoroutineTests.kt
@@ -18,7 +18,7 @@ package io.realm
 
 import io.realm.internal.platform.singleThreadDispatcher
 import io.realm.util.NsQueueDispatcher
-import io.realm.util.PlatformUtils
+import io.realm.test.platform.PlatformUtils
 import io.realm.util.Utils.printlntid
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
